Problema al validar las contraseñas

+++ backend/src/main/java/com/pe/inventoryapp/backend/user/controller/UserController.java
@@ -7,11 +7,14 @@ import com.pe.inventoryapp.backend.auth.models.UserPrincipal;
 import com.pe.inventoryapp.backend.auth.service.AuthService;
 import com.pe.inventoryapp.backend.common.response.Response;
 import com.pe.inventoryapp.backend.common.service.ResponseService;
+import com.pe.inventoryapp.backend.user.model.request.PasswordRequest;
+import com.pe.inventoryapp.backend.user.model.request.ProfileRequest;
 import com.pe.inventoryapp.backend.user.model.response.DetailUserResponse;
 import com.pe.inventoryapp.backend.user.service.UserService;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
+import jakarta.validation.Valid;
 
 import static com.pe.inventoryapp.backend.security.config.TokenJwtConfig.SECRET_KEY;
 
@@ -26,10 +29,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 @RestController
 @RequestMapping("/api/users")
@@ -70,13 +76,76 @@ public class UserController {
     return ResponseEntity.ok(user);
   }
 
+  @PutMapping("profile")
+  public ResponseEntity<Response> updateProfile(@RequestHeader("Authorization") String header,
+      @Valid @RequestBody ProfileRequest profileRequest, BindingResult result) {
+    Long id = authService.extracIdFromClaims(header);
+
+    if (userService.findById(id).isEmpty()) {
+      return ResponseEntity.status(400)
+          .body(responseService.writeAResponse("error", "El usuario no existe"));
+    }
+
+    if (result.hasErrors()) {
+      return ResponseEntity.status(400).body(responseService.writeAResponse("error", "Complete los campos si o si"));
+    }
+
+    String message = userService.updateProfile(id, profileRequest);
+    return ResponseEntity.status(200).body(responseService.writeAResponse("success", message));
+  }
+
+  @PutMapping("update-password")
+  public ResponseEntity<Response> updatePassword(@RequestHeader("Authorization") String header,
+      @Valid @RequestBody PasswordRequest passwordRequest, BindingResult result) {
+    Long id = authService.extracIdFromClaims(header);
+
+    if (userService.findById(id).isEmpty()) {
+      return ResponseEntity.status(400)
+          .body(responseService.writeAResponse("error", "El usuario no existe"));
+    }
+
+    if (result.hasErrors()) {
+      return ResponseEntity.status(400).body(responseService.writeAResponse("error", "Complete los campos si o si"));
+    }
+
+    if (userService.validatePassword(id, passwordRequest)) {
+
+      // TODO : CORREGIR AQUI, LAS CONTRASEÑAS NO COINCIDEN
+      if (passwordRequest.getNewPassword() != passwordRequest.getConfirmPassword()) {
+        return ResponseEntity.status(400)
+            .body(responseService.writeAResponse("error", "Confirma tu contraseña, parece que no es la misma"));
+      } else {
+        String message = userService.updatePassword(id, passwordRequest);
+        return ResponseEntity.status(200).body(responseService.writeAResponse("success", message));
+
+      }
+
+    } else {
+      return ResponseEntity.status(400)
+          .body(responseService.writeAResponse("error", "Intente de nuevo, no es la contraseña anterior"));
+    }
+
+  }
+
   @DeleteMapping("/{id}")
   public ResponseEntity<Response> deleteUser(@PathVariable Long id) {
-    String message = userService.remove(id);
+
+    if (userService.findById(id).isEmpty()) {
+      return ResponseEntity.status(400)
+          .body(responseService.writeAResponse("error", "El usuario no existe"));
+    }
+
+    if (id == 1) {
+      return ResponseEntity.status(400)
+          .body(responseService.writeAResponse("error", "Este usuario no se puede eliminar"));
+
+    }
+
+    userService.remove(id);
     // return ResponseEntity.ok("Se ha eliminado el usuario");
 
     return ResponseEntity.status(HttpStatus.CREATED)
-        .body(responseService.writeAResponse("success", message));
+        .body(responseService.writeAResponse("success", "Usuario eliminado"));
 
   }
